# apps/manager/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@sysarmor.com
    name: SysArmor Team
    url: https://sysarmor.com/support
  description: SysArmor EDR 系统的控制平面服务 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sysarmor.com/terms
  title: SysArmor Manager API
  version: "1.0"
paths:
...
  /events/collectors/{collector_id}:
    get:
      consumes:
      - application/json
      description: 根据 Collector ID 查询其相关的所有事件数据
      parameters:
      - description: Collector ID
        in: path
        name: collector_id
        required: true
        type: string
      - description: 事件类型过滤
        in: query
        name: event_type
        type: string
      - description: 返回数量限制，默认100
        in: query
        name: limit
        type: integer
      - description: 是否只获取最新事件
        in: query
        name: latest
        type: boolean
      - description: 开始时间 (RFC3339格式)
        in: query
        name: from_time
        type: string
      - description: 结束时间 (RFC3339格式)
        in: query
        name: to_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Collector 事件查询结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 查询特定 Collector 的事件
      tags:
      - events
  /events/query:
    get:
      consumes:
      - application/json
      description: 根据 Topic 和其他条件查询事件数据
      parameters:
      - description: Kafka Topic 名称
        in: query
        name: topic
        required: true
        type: string
      - description: Collector ID 过滤
        in: query
        name: collector_id
        type: string
      - description: 事件类型过滤
        in: query
        name: event_type
        type: string
      - description: 返回数量限制，默认100
        in: query
        name: limit
        type: integer
      - description: 是否只获取最新事件
        in: query
        name: latest
        type: boolean
      - description: 开始时间 (RFC3339格式)
        in: query
        name: from_time
        type: string
      - description: 结束时间 (RFC3339格式)
        in: query
        name: to_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 事件查询结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 查询事件
      tags:
      - events
  /events/topics:
    get:
      consumes:
      - application/json
      description: 获取所有可用的 Kafka Topic 列表，包括 Collector 相关和其他 Topic
      produces:
      - application/json
      responses:
        "200":
          description: Topic 列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 列出所有 Kafka Topics
      tags:
      - events
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"